Middleware works but users.test.ts doesn't

+++ tests/posts.test.ts
@@ -7,12 +7,20 @@ import { Express } from "express";
 let testApp: Express;
 
 beforeAll(async () => {
+    
+    // Close old connections 
+    await mongoose.connection.close();
+    if (mongoose.connection.readyState === 0) {
+        const dbUri = process.env.TEST_MONGO_URL || "";
+        console.log(`Connecting to MongoDB at ${dbUri}`);
+
+        // Connect to MongoDB
+        await mongoose.connect(process.env.TEST_MONGO_URL || "");
+        await Post.deleteMany();
+        console.log('MongoDB connection established');
+    }
+
     testApp = app;
-
-    // Connect to MongoDB
-    await mongoose.connect(process.env.TEST_MONGO_URL || "");
-    await Post.deleteMany();
-
 });
 
 afterAll(async () => {
